FIX: VC++ warnings
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3884 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Segmentation/cmtkLabelCombinationLocalBinaryShapeBasedAveraging.cxx
@@ -88,7 +88,7 @@ cmtk::LabelCombinationLocalBinaryShapeBasedAveraging::ComputeResultForRegion( co
   std::vector<short> labels( nAtlases );  
   std::vector<Types::DataItem> weights( nAtlases );  
   std::vector<size_t> bestPatchOffset( nAtlases );
-  std::vector<float> distances( nAtlases );
+  std::vector<double> distances( nAtlases );
 
   for ( RegionIndexIterator<TargetRegionType> it( region ); it != it.end(); ++it )
     {
@@ -116,19 +116,19 @@ cmtk::LabelCombinationLocalBinaryShapeBasedAveraging::ComputeResultForRegion( co
       std::sort( distances.begin(), distances.begin()+nn );
 
       // determine 1st and 3rd quartile values
-      const float Q1 = distances[static_cast<size_t>( 0.25 * nn )];
-      const float Q3 = distances[static_cast<size_t>( 0.75 * nn )];
+      const double Q1 = distances[static_cast<size_t>( 0.25 * nn )];
+      const double Q3 = distances[static_cast<size_t>( 0.75 * nn )];
       
       // compute thresholds from quartiles and inter-quartile range
-      const float lThresh = Q1 - 1.5 * (Q3-Q1);
-      const float uThresh = Q3 + 1.5 * (Q3-Q1);
+      const double lThresh = Q1 - 1.5 * (Q3-Q1);
+      const double uThresh = Q3 + 1.5 * (Q3-Q1);
       
       // mark as invalid those atlases with values outside the "inlier" range
       for ( size_t n = 0; n < nAtlases; ++n )
 	{
 	if ( valid[n] )
 	  {
-	  const float d = this->m_AtlasDMaps[n]->GetDataAt( i );
+	  const double d = this->m_AtlasDMaps[n]->GetDataAt( i );
 	  if ( (d < lThresh) || (d > uThresh) )
 	    valid[n] = false;
 	  }
